FE-1446 - Re-organize tests
