+++ src/index.js
@@ -7,7 +7,7 @@ const {
 const { getSecurityVulnerabilities, postSlackMsg } = require('./api.js');
 const { formatVulnerabilityAlerts, getIntroMsg } = require('./helpers.js');
 
-async function start() {
+async function main() {
   const { data } = await getSecurityVulnerabilities({
     githubRepo: GITHUB_REPOSITORY,
     githubToken: GITHUB_TOKEN,
@@ -17,7 +17,10 @@ async function start() {
   if (Boolean(vulnerabilityAlerts) && vulnerabilityAlerts.length === 0)
     return console.log('No security vulnerabilities found.');
 
-  const introMsg = getIntroMsg(vulnerabilityAlerts.length);
+  const introMsg = getIntroMsg({
+    numberOfVulnerabilities: vulnerabilityAlerts.length,
+    githubRepo: GITHUB_REPOSITORY,
+  });
 
   await postSlackMsg({
     slackChannel: SLACK_CHANNEL,
@@ -67,4 +70,9 @@ async function start() {
   });
 }
 
-start();
+main();
+
+// Exported for testing purposes
+module.exports = {
+  main,
+};
